+++ controllers/doguExportManager_test.go
@@ -2,11 +2,13 @@ package controllers
 
 import (
 	"context"
+	"errors"
 	"github.com/cloudogu/cesapp-lib/core"
 	doguv2 "github.com/cloudogu/k8s-dogu-lib/v2/api/v2"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
+	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"testing"
@@ -25,11 +27,18 @@ func Test_doguExportManager_UpdateExportMode(t *testing.T) {
 			{Status: corev1.PodStatus{ContainerStatuses: []corev1.ContainerStatus{{Name: "myDogu-exporter", Ready: false}}}},
 		}}
 
+		deploymentList := &appsv1.DeploymentList{Items: []appsv1.Deployment{
+			{Spec: appsv1.DeploymentSpec{Template: corev1.PodTemplateSpec{Spec: corev1.PodSpec{Containers: nil}}}},
+		}}
+
 		mockDoguClient := newMockDoguInterface(t)
 		mockDoguClient.EXPECT().UpdateStatusWithRetry(testCtx, doguResource, mock.Anything, metav1.UpdateOptions{}).Return(nil, nil)
 
 		mockPodClient := newMockPodInterface(t)
 		mockPodClient.EXPECT().List(testCtx, metav1.ListOptions{LabelSelector: "dogu.name=myDogu"}).Return(podList, nil)
+
+		mockDeploymentClient := newMockDeploymentInterface(t)
+		mockDeploymentClient.EXPECT().List(testCtx, metav1.ListOptions{LabelSelector: "dogu.name=myDogu"}).Return(deploymentList, nil)
 
 		mockDoguFetcher := newMockLocalDoguFetcher(t)
 		mockDoguFetcher.EXPECT().FetchInstalled(testCtx, doguResource.GetSimpleDoguName()).Return(dogu, nil)
@@ -40,6 +49,7 @@ func Test_doguExportManager_UpdateExportMode(t *testing.T) {
 		dem := &doguExportManager{
 			doguClient:       mockDoguClient,
 			podClient:        mockPodClient,
+			deploymentClient: mockDeploymentClient,
 			doguFetcher:      mockDoguFetcher,
 			resourceUpserter: mockUpserter,
 		}
@@ -53,25 +63,19 @@ func Test_doguExportManager_UpdateExportMode(t *testing.T) {
 		assert.True(t, err.(exportModeNotYetChangedError).Requeue())
 	})
 
-	t.Run("should update deployment when could not get current state of export-mode", func(t *testing.T) {
+	t.Run("reconcile when could not get current state of export-mode", func(t *testing.T) {
 		doguResource := &doguv2.Dogu{
 			ObjectMeta: metav1.ObjectMeta{Name: "myDogu"},
 			Spec:       doguv2.DoguSpec{ExportMode: true},
 		}
 
-		dogu := &core.Dogu{}
-
 		mockDoguClient := newMockDoguInterface(t)
-		mockDoguClient.EXPECT().UpdateStatusWithRetry(testCtx, doguResource, mock.Anything, metav1.UpdateOptions{}).Return(nil, nil)
 
 		mockPodClient := newMockPodInterface(t)
 		mockPodClient.EXPECT().List(testCtx, metav1.ListOptions{LabelSelector: "dogu.name=myDogu"}).Return(nil, assert.AnError)
 
 		mockDoguFetcher := newMockLocalDoguFetcher(t)
-		mockDoguFetcher.EXPECT().FetchInstalled(testCtx, doguResource.GetSimpleDoguName()).Return(dogu, nil)
-
 		mockUpserter := newMockResourceUpserter(t)
-		mockUpserter.EXPECT().UpsertDoguDeployment(testCtx, doguResource, dogu, mock.Anything).Return(nil, nil)
 
 		dem := &doguExportManager{
 			doguClient:       mockDoguClient,
@@ -86,6 +90,9 @@ func Test_doguExportManager_UpdateExportMode(t *testing.T) {
 		assert.ErrorContains(t, err, "error while changing the export-mode of dogu \"myDogu\": failed to check if deployment is in export-mode dogu \"myDogu\": failed to get pods of deployment \"/myDogu\":")
 		assert.Equal(t, requeueWaitTimeout, err.(exportModeNotYetChangedError).GetRequeueTime())
 		assert.True(t, err.(exportModeNotYetChangedError).Requeue())
+
+		mockDoguFetcher.AssertNotCalled(t, "FetchInstalled")
+		mockUpserter.AssertNotCalled(t, "UpsertDoguDeployment")
 	})
 
 	t.Run("should fail to update deployment when export-mode changes on error updating status", func(t *testing.T) {
@@ -98,21 +105,32 @@ func Test_doguExportManager_UpdateExportMode(t *testing.T) {
 			{Status: corev1.PodStatus{ContainerStatuses: []corev1.ContainerStatus{{Name: "myDogu-exporter", Ready: false}}}},
 		}}
 
+		deploymentList := &appsv1.DeploymentList{Items: []appsv1.Deployment{
+			{Spec: appsv1.DeploymentSpec{Template: corev1.PodTemplateSpec{Spec: corev1.PodSpec{Containers: nil}}}},
+		}}
+
 		mockDoguClient := newMockDoguInterface(t)
 		mockDoguClient.EXPECT().UpdateStatusWithRetry(testCtx, doguResource, mock.Anything, metav1.UpdateOptions{}).Return(nil, assert.AnError)
 
 		mockPodClient := newMockPodInterface(t)
 		mockPodClient.EXPECT().List(testCtx, metav1.ListOptions{LabelSelector: "dogu.name=myDogu"}).Return(podList, nil)
 
+		mockDeploymentClient := newMockDeploymentInterface(t)
+		mockDeploymentClient.EXPECT().List(testCtx, metav1.ListOptions{LabelSelector: "dogu.name=myDogu"}).Return(deploymentList, nil)
+
 		dem := &doguExportManager{
-			doguClient: mockDoguClient,
-			podClient:  mockPodClient,
+			doguClient:       mockDoguClient,
+			podClient:        mockPodClient,
+			deploymentClient: mockDeploymentClient,
 		}
 
 		err := dem.UpdateExportMode(testCtx, doguResource)
 		require.Error(t, err)
 
-		assert.ErrorIs(t, err, assert.AnError)
+		var oErr exportModeNotYetChangedError
+		assert.True(t, errors.As(err, &oErr))
+		assert.True(t, oErr.Requeue())
+		assert.ErrorIs(t, oErr.err, assert.AnError)
 	})
 
 	t.Run("should not update deployment when export-mode already in desired state", func(t *testing.T) {
@@ -124,7 +142,9 @@ func Test_doguExportManager_UpdateExportMode(t *testing.T) {
 		//dogu := &core.Dogu{}
 
 		podList := &corev1.PodList{Items: []corev1.Pod{
-			{Status: corev1.PodStatus{ContainerStatuses: []corev1.ContainerStatus{{Name: "myDogu-exporter", Ready: true}}}},
+			{Status: corev1.PodStatus{
+				Phase:             corev1.PodRunning,
+				ContainerStatuses: []corev1.ContainerStatus{{Name: "myDogu-exporter", Ready: true}}}},
 		}}
 
 		mockDoguClient := newMockDoguInterface(t)
@@ -141,6 +161,41 @@ func Test_doguExportManager_UpdateExportMode(t *testing.T) {
 		err := dem.UpdateExportMode(testCtx, doguResource)
 		require.NoError(t, err)
 	})
+
+	t.Run("should not update deployment when export-mode already set in deployment spec", func(t *testing.T) {
+		doguResource := &doguv2.Dogu{
+			ObjectMeta: metav1.ObjectMeta{Name: "myDogu"},
+			Spec:       doguv2.DoguSpec{ExportMode: true},
+		}
+
+		podList := &corev1.PodList{Items: []corev1.Pod{
+			{Status: corev1.PodStatus{ContainerStatuses: []corev1.ContainerStatus{{Name: "myDogu-exporter", Ready: false}}}},
+		}}
+
+		deploymentList := &appsv1.DeploymentList{Items: []appsv1.Deployment{
+			{Spec: appsv1.DeploymentSpec{Template: corev1.PodTemplateSpec{Spec: corev1.PodSpec{Containers: []corev1.Container{{Name: "myDogu-exporter"}}}}}},
+		}}
+
+		mockPodClient := newMockPodInterface(t)
+		mockPodClient.EXPECT().List(testCtx, metav1.ListOptions{LabelSelector: "dogu.name=myDogu"}).Return(podList, nil)
+
+		mockDeploymentClient := newMockDeploymentInterface(t)
+		mockDeploymentClient.EXPECT().List(testCtx, metav1.ListOptions{LabelSelector: "dogu.name=myDogu"}).Return(deploymentList, nil)
+
+		dem := &doguExportManager{
+			podClient:        mockPodClient,
+			deploymentClient: mockDeploymentClient,
+		}
+
+		err := dem.UpdateExportMode(testCtx, doguResource)
+
+		require.Error(t, err)
+		require.ErrorIs(t, err, exportModeNotYetChangedError{doguName: "myDogu", desiredExportModeState: true})
+		assert.Equal(t, "the export-mode of dogu \"myDogu\" has not yet been changed to its desired state: true", err.Error())
+		assert.Equal(t, requeueWaitTimeout, err.(exportModeNotYetChangedError).GetRequeueTime())
+		assert.True(t, err.(exportModeNotYetChangedError).Requeue())
+		assert.Nil(t, err.(exportModeNotYetChangedError).err)
+	})
 }
 
 func Test_doguExportManager_shouldUpdateExportMode(t *testing.T) {
@@ -151,7 +206,9 @@ func Test_doguExportManager_shouldUpdateExportMode(t *testing.T) {
 		}
 
 		podList := &corev1.PodList{Items: []corev1.Pod{
-			{Status: corev1.PodStatus{ContainerStatuses: []corev1.ContainerStatus{{Name: "myDogu-exporter", Ready: false}}}},
+			{Status: corev1.PodStatus{
+				Phase:             corev1.PodRunning,
+				ContainerStatuses: []corev1.ContainerStatus{{Name: "myDogu-exporter", Ready: false}}}},
 		}}
 
 		mockPodClient := newMockPodInterface(t)
@@ -173,7 +230,9 @@ func Test_doguExportManager_shouldUpdateExportMode(t *testing.T) {
 		}
 
 		podList := &corev1.PodList{Items: []corev1.Pod{
-			{Status: corev1.PodStatus{ContainerStatuses: []corev1.ContainerStatus{{Name: "myDogu-exporter", Ready: true}}}},
+			{Status: corev1.PodStatus{
+				Phase:             corev1.PodRunning,
+				ContainerStatuses: []corev1.ContainerStatus{{Name: "myDogu-exporter", Ready: true}}}},
 		}}
 
 		mockPodClient := newMockPodInterface(t)
@@ -224,9 +283,18 @@ func Test_doguExportManager_updateExportMode(t *testing.T) {
 		mockDoguFetcher := newMockLocalDoguFetcher(t)
 		mockDoguFetcher.EXPECT().FetchInstalled(testCtx, doguResource.GetSimpleDoguName()).Return(dogu, assert.AnError)
 
+		deploymentList := &appsv1.DeploymentList{Items: []appsv1.Deployment{
+			//{Spec: appsv1.DeploymentSpec{Template: corev1.PodTemplateSpec{Spec: corev1.PodSpec{Containers: []corev1.Container{{Name: "myDogu-exporter"}}}}}},
+			{Spec: appsv1.DeploymentSpec{Template: corev1.PodTemplateSpec{Spec: corev1.PodSpec{Containers: nil}}}},
+		}}
+
+		mockDeploymentClient := newMockDeploymentInterface(t)
+		mockDeploymentClient.EXPECT().List(testCtx, metav1.ListOptions{LabelSelector: "dogu.name=myDogu"}).Return(deploymentList, nil)
+
 		dem := &doguExportManager{
-			doguClient:  mockDoguClient,
-			doguFetcher: mockDoguFetcher,
+			doguClient:       mockDoguClient,
+			doguFetcher:      mockDoguFetcher,
+			deploymentClient: mockDeploymentClient,
 		}
 
 		err := dem.updateExportMode(testCtx, doguResource)
@@ -250,12 +318,21 @@ func Test_doguExportManager_updateExportMode(t *testing.T) {
 		mockDoguFetcher := newMockLocalDoguFetcher(t)
 		mockDoguFetcher.EXPECT().FetchInstalled(testCtx, doguResource.GetSimpleDoguName()).Return(dogu, nil)
 
+		deploymentList := &appsv1.DeploymentList{Items: []appsv1.Deployment{
+			//{Spec: appsv1.DeploymentSpec{Template: corev1.PodTemplateSpec{Spec: corev1.PodSpec{Containers: []corev1.Container{{Name: "myDogu-exporter"}}}}}},
+			{Spec: appsv1.DeploymentSpec{Template: corev1.PodTemplateSpec{Spec: corev1.PodSpec{Containers: nil}}}},
+		}}
+
+		mockDeploymentClient := newMockDeploymentInterface(t)
+		mockDeploymentClient.EXPECT().List(testCtx, metav1.ListOptions{LabelSelector: "dogu.name=myDogu"}).Return(deploymentList, nil)
+
 		mockUpserter := newMockResourceUpserter(t)
 		mockUpserter.EXPECT().UpsertDoguDeployment(testCtx, doguResource, dogu, mock.Anything).Return(nil, assert.AnError)
 
 		dem := &doguExportManager{
 			doguClient:       mockDoguClient,
 			doguFetcher:      mockDoguFetcher,
+			deploymentClient: mockDeploymentClient,
 			resourceUpserter: mockUpserter,
 		}